backend-101/lesson-5/pyproject.toml
[tool.poetry]
name = "back-end-101"
version = "0.1.0"
description = ""
authors = ["Anuar Kopzhassarov <chehve@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
